// src/lib/api/generated/models/createRecommendationRuleDtoTargetType.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Beauty Pack Recommendation API
 * Backend API for beauty-pack personalization, recommendations, COD orders, and administration
 * OpenAPI spec version: 1.0.0
 */

export type CreateRecommendationRuleDtoTargetType = typeof CreateRecommendationRuleDtoTargetType[keyof typeof CreateRecommendationRuleDtoTargetType];


export const CreateRecommendationRuleDtoTargetType = {
  PACK: 'PACK',
  PRODUCT: 'PRODUCT',
  REFERENCE: 'REFERENCE',
} as const;