// src/lib/api/generated/models/createRecommendationRuleDtoConditionType.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Beauty Pack Recommendation API
 * Backend API for beauty-pack personalization, recommendations, COD orders, and administration
 * OpenAPI spec version: 1.0.0
 */

export type CreateRecommendationRuleDtoConditionType = typeof CreateRecommendationRuleDtoConditionType[keyof typeof CreateRecommendationRuleDtoConditionType];


export const CreateRecommendationRuleDtoConditionType = {
  MUST_MATCH: 'MUST_MATCH',
  SHOULD_MATCH: 'SHOULD_MATCH',
  EXCLUDE_IF_MATCH: 'EXCLUDE_IF_MATCH',
} as const;